feat: implement image fetching and configuration for dynamic routes
