+++ packages/web/nuxt.config.ts
@@ -1,29 +1,7 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 
-import { fetchArticleRoutes } from "./scripts/fetchArticleRoutes";
-import { fetchImages } from "./scripts/fetchImages";
-
-export const STAGE_NAME = process.env.STAGE_NAME ?? "dev";
-
-if (!["dev", "stg", "prod"].includes(STAGE_NAME)) {
-  throw new Error("STAGE_NAME is not valid.");
-}
-
-export const ENDPOINT =
-  STAGE_NAME === "prod"
-    ? `https://www.46ki75.com`
-    : `https://${STAGE_NAME}-www.46ki75.com`;
-
-const GTAG =
-  STAGE_NAME === "prod"
-    ? "G-TW1BVM24YT"
-    : STAGE_NAME === "stg"
-    ? "G-Q7K53RM4VC"
-    : "G-85QSG3WH5F";
-
-const routes = await fetchArticleRoutes(ENDPOINT);
-
-await fetchImages();
+import { ENDPOINT, GTAG, STAGE_NAME } from "./scripts/fetchConfig";
+import { fetchPrerenderRoutes } from "./scripts/fetchRoutes";
 
 export default defineNuxtConfig({
   compatibilityDate: "2024-11-01",
@@ -48,14 +26,7 @@ export default defineNuxtConfig({
   },
   nitro: {
     prerender: {
-      routes: [
-        ...routes,
-        "/",
-        "/about",
-        "/blog",
-        "/blog/article",
-        "/blog/search",
-      ],
+      routes: await fetchPrerenderRoutes(ENDPOINT),
       crawlLinks: false,
       concurrency: 10,
     },
