+++ packages/web/package.json
@@ -3,8 +3,9 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "nuxt build",
+    "predev": "tsx ./scripts/prepare.ts",
     "dev": "nuxt dev",
+    "pregenerate": "tsx ./scripts/prepare.ts",
     "generate": "nuxt generate",
     "preview": "nuxt preview",
     "postinstall": "nuxt prepare",
@@ -38,6 +39,7 @@
     "stylelint": "^16.17.0",
     "stylelint-config-recommended-vue": "^1.6.0",
     "stylelint-config-standard-scss": "^14.0.0",
+    "tsx": "^4.19.3",
     "typescript-eslint": "^8.28.0"
   }
 }
