+++ packages/web/scripts/prepare.ts
@@ -1,0 +1,3 @@
+import { fetchImages } from "./fetchImages";
+
+await fetchImages();
